Add more functionality to test app
Add user detail and simplify loader component. Create a popover component
to be used as the navigation eventually.

Add eslinter to the project and bring in vue-eslint-parser as the
engine and add parserOptions according to documentation. User beforeRouteEnter
to fetch the users before continuing route to ensure users
are there for the detail screen.

Move some SCSS in to the components and scope it. Create an alias in
webpack to resolve the variables file so it's easy to import in
to any file.